Added demo for blitting

+++ src/arch/x86_64/drivers/vbe_video.c
@@ -12,8 +12,8 @@ void vbe_swap_buffers()
 
 void vbe_video_init(multiboot_tag_framebuffer_common_t* common)
 {
-  video_driver.width = common->framebuffer_width;
-  video_driver.height = common->framebuffer_height;
+  video_driver.width = (int32_t)common->framebuffer_width;
+  video_driver.height = (int32_t)common->framebuffer_height;
   video_driver.native_buffer = common->framebuffer_addr;
   video_driver.buffer = (uint32_t*)malloc(
     common->framebuffer_width * common->framebuffer_height * sizeof(uint32_t));
